mbt-core/pyproject.toml
[project]
name = "mbt-core"
version = "0.1.0"
description = "MBT (Model Build Tool) - Declarative ML pipeline framework"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer[all]>=0.12.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "jinja2>=3.1.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "joblib>=1.3.0",
    "rich>=13.0.0",  # For beautiful CLI output
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
mbt = "mbt.cli:app"

[project.entry-points."mbt.storage"]
local = "mbt.builtins.local_storage:LocalStoragePlugin"

[project.entry-points."mbt.data_connectors"]
local_file = "mbt.builtins.local_connector:LocalFileConnector"

[project.entry-points."mbt.executors"]
local = "mbt.builtins.local_executor:LocalExecutor"

[project.entry-points."mbt.secrets"]
env = "mbt.builtins.env_secrets:EnvSecretsProvider"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
